[Overflow-66] [Markup] Teams Question Tab
typescript error fixes
